Move interaction with db to separate file. To make easier in future a possible change ORM and have related functions in one place
