+++ src/app/auth/login/actions.ts
@@ -1,17 +1,15 @@
 "use server"
 
 import { z } from "zod"
-import { createSession, deleteSession } from "../lib/session"
 import { redirect } from "next/navigation"
 import * as bcrypt from "bcrypt"
-import { PrismaClient } from "@prisma/client"
+import { createSession, deleteSession } from "@/lib/session"
+import { findUserByEmail } from "@/lib/users"
 
 const loginSchema = z.object({
   email: z.string().email({ message: "Invalid email address" }).trim(),
   password: z.string().min(6, { message: "Password must be at least 6 characters" }).trim(),
 })
-
-const prisma = new PrismaClient()
 
 export async function login(prevState: any, formData: FormData) {
   // Validate input
@@ -25,10 +23,8 @@ export async function login(prevState: any, formData: FormData) {
 
   const { email, password } = result.data
 
-  // Check for user in db and password
-  const user = await prisma.user.findUnique({
-    where: { email },
-  })
+  // Check for user in db and check password
+  const user = await findUserByEmail(email)
 
   if (!user) {
     return {
