+++ src/app/auth/signup/actions.ts
@@ -1,17 +1,15 @@
 "use server"
 
 import { z } from "zod"
-import { createSession } from "../lib/session"
 import { redirect } from "next/navigation"
-import { PrismaClient } from "@prisma/client"
 import * as bcrypt from "bcrypt"
+import { createSession } from "@/lib/session"
+import { createUser, findUserByEmail } from "@/lib/users"
 
 const loginSchema = z.object({
   email: z.string().email({ message: "Invalid email address" }).trim(),
   password: z.string().min(6, { message: "Password must be at least 6 characters" }).trim(),
 })
-
-const prisma = new PrismaClient()
 
 export async function signup(prevState: any, formData: FormData) {
   // Validate input
@@ -26,7 +24,7 @@ export async function signup(prevState: any, formData: FormData) {
   const { email, password } = result.data
 
   // Check user in DB
-  const emailAlreadyExists = await prisma.user.findUnique({ where: { email } })
+  const emailAlreadyExists = await findUserByEmail(email)
 
   if (emailAlreadyExists) {
     return {
@@ -39,9 +37,7 @@ export async function signup(prevState: any, formData: FormData) {
   // Create user and session
   const hashedPassword = await bcrypt.hash(password, 10)
 
-  const newUser = await prisma.user.create({
-    data: { email, password: hashedPassword },
-  })
+  const newUser = await createUser(email, hashedPassword)
 
   await createSession(email)
